check for family in findDegeneracies
modified the findDegeneracies to fix bug that resulted from
finding the same reactions if forward and backward directions are
flipped (like in beta scission reactions). At the same time
this commit requires the same family for the reaction to go through
adding a requirement for the same template can be done later after
multiple transition states are supported

squash! check for family and template in findDegeneracies