make reaction direction deterministic
model generation sometimes flips the direction of a reaction.
This seems to be caused by floating point errors in calculating
the deltaGrxn. This ensures the same direction is chosen as long
as deltaG is within 0.01 microJ/mol which is 100x the floating point
error obeserved on h abstraction from propane by butyl radical.